Refactor catalog and settings structure
