+++ backend/app/routers/books.py
@@ -1,7 +1,6 @@
 """Book CRUD, search, chapter listing, and download endpoints."""
 
 import logging
-from datetime import datetime, timedelta, timezone
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -12,121 +11,14 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from .. import crud, epub_editor, models, schemas
 from ..config import LIBRARY_PATH
 from ..database import get_db
+from ..services.catalog import build_book_catalog, normalize_genre_tags
+from ..services.chapter_history import build_chapter_update_history
 from ..services.library_paths import remove_empty_parent_dirs
 from ..services.metadata_jobs import queue_metadata_sync_job
 
 logger = logging.getLogger(__name__)
 
 router = APIRouter()
-
-WORDS_PER_MONTH_WEEKS = 52 / 12
-CATCH_UP_SYNC_MIN_CHAPTERS = 5
-CATCH_UP_SYNC_MIN_WORDS = 20_000
-
-
-def _as_utc(dt: datetime) -> datetime:
-    if dt.tzinfo is None:
-        return dt.replace(tzinfo=timezone.utc)
-    return dt.astimezone(timezone.utc)
-
-
-def _chapter_delta(log: models.BookLog) -> int:
-    if log.previous_chapter_count is None or log.new_chapter_count is None:
-        return 0
-    return max(log.new_chapter_count - log.previous_chapter_count, 0)
-
-
-def _is_chapter_growth_log(log: models.BookLog) -> bool:
-    return (
-        log.entry_type == "updated"
-        and log.previous_chapter_count is not None
-        and log.previous_chapter_count > 0
-        and _chapter_delta(log) > 0
-    )
-
-
-def _is_initial_sync_log(log: models.BookLog) -> bool:
-    return log.entry_type == "added" and (log.new_chapter_count or 0) > 0
-
-
-def _build_chapter_update_history(
-    book_id: int,
-    logs: list[models.BookLog],
-    now: datetime | None = None,
-) -> schemas.BookChapterUpdateHistory:
-    points: list[schemas.BookChapterUpdateHistoryPoint] = []
-    seen_initial_sync = False
-    seen_post_initial_growth = False
-    for log in logs:
-        is_initial_sync = _is_initial_sync_log(log)
-        included_in_stats = _is_chapter_growth_log(log)
-        if not is_initial_sync and not included_in_stats:
-            continue
-        chapters_added = (log.new_chapter_count or 0) if is_initial_sync else _chapter_delta(log)
-        words_added = max(log.words_added or 0, 0)
-        is_catch_up_sync = (
-            included_in_stats
-            and seen_initial_sync
-            and not seen_post_initial_growth
-            and (chapters_added >= CATCH_UP_SYNC_MIN_CHAPTERS or words_added >= CATCH_UP_SYNC_MIN_WORDS)
-        )
-        included_in_stats = included_in_stats and not is_catch_up_sync
-        points.append(
-            schemas.BookChapterUpdateHistoryPoint(
-                id=log.id,
-                timestamp=log.timestamp,
-                entry_type=log.entry_type,
-                previous_chapter_count=log.previous_chapter_count,
-                new_chapter_count=log.new_chapter_count,
-                chapters_added=chapters_added,
-                words_added=words_added,
-                average_words_per_chapter=words_added / chapters_added if chapters_added else None,
-                included_in_stats=included_in_stats,
-                is_initial_sync=is_initial_sync,
-                is_catch_up_sync=is_catch_up_sync,
-            )
-        )
-        if is_initial_sync:
-            seen_initial_sync = True
-        elif _is_chapter_growth_log(log):
-            seen_post_initial_growth = True
-
-    stats_points = [point for point in points if point.included_in_stats]
-    total_words_added = sum(point.words_added for point in stats_points)
-    total_chapters_added = sum(point.chapters_added for point in stats_points)
-    average_words_per_week = None
-    average_words_per_month = None
-    average_days_between_updates = None
-    predicted_next_update_at = None
-    last_update_at = stats_points[-1].timestamp if stats_points else None
-
-    if stats_points:
-        now_utc = _as_utc(now or datetime.now(timezone.utc))
-        first_update_at = _as_utc(stats_points[0].timestamp)
-        elapsed_days = max((now_utc - first_update_at).total_seconds() / 86400, 1)
-        average_words_per_week = total_words_added / (elapsed_days / 7)
-        average_words_per_month = average_words_per_week * WORDS_PER_MONTH_WEEKS
-
-    if len(stats_points) >= 2:
-        timestamps = [_as_utc(point.timestamp) for point in stats_points]
-        intervals = [(current - previous).total_seconds() / 86400 for previous, current in zip(timestamps, timestamps[1:])]
-        average_days_between_updates = sum(intervals) / len(intervals)
-        predicted_next_update_at = timestamps[-1] + timedelta(days=average_days_between_updates)
-
-    return schemas.BookChapterUpdateHistory(
-        book_id=book_id,
-        history=points,
-        summary=schemas.BookChapterUpdateHistorySummary(
-            total_update_events=len(stats_points),
-            total_chapters_added=total_chapters_added,
-            total_words_added=total_words_added,
-            average_words_per_week=average_words_per_week,
-            average_words_per_month=average_words_per_month,
-            average_days_between_updates=average_days_between_updates,
-            predicted_next_update_at=predicted_next_update_at,
-            last_update_at=last_update_at,
-        ),
-    )
 
 
 def _remove_book_files(book: models.Book) -> list[str]:
@@ -164,52 +56,6 @@ def _book_cleanup_preview(book: models.Book) -> dict[str, Any]:
     }
 
 
-def _normalize_genre_tags(tags: list[str]) -> list[str]:
-    normalized: list[str] = []
-    seen: set[str] = set()
-    for raw_tag in tags:
-        cleaned = raw_tag.strip()
-        if not cleaned:
-            continue
-        folded = cleaned.casefold()
-        if folded in seen:
-            continue
-        seen.add(folded)
-        normalized.append(cleaned)
-    return sorted(normalized, key=str.casefold)
-
-
-def _effective_genre_tags(book: models.Book, series_user_genre_tags: list[str] | None = None) -> list[str]:
-    return _normalize_genre_tags(
-        [
-            *(series_user_genre_tags or []),
-            *(book.user_genre_tags or []),
-            *(book.genre_tags or []),
-        ]
-    )
-
-
-async def _series_metadata_map(
-    db: AsyncSession,
-    books: list[models.Book],
-) -> dict[str, models.SeriesMetadata]:
-    series_names = sorted({book.series for book in books if book.series})
-    return await crud.get_series_metadata_for_names(db, series_names)
-
-
-def _serialize_catalog_book(
-    book: models.Book,
-    *,
-    series_user_genre_tags: list[str] | None = None,
-    effective_series_genre_tags: list[str] | None = None,
-) -> schemas.BookCatalogEntry:
-    payload = schemas.BookCatalogEntry.model_validate(book).model_dump()
-    payload["series_user_genre_tags"] = series_user_genre_tags or []
-    payload["effective_genre_tags"] = _effective_genre_tags(book, series_user_genre_tags)
-    payload["effective_series_genre_tags"] = effective_series_genre_tags or []
-    return schemas.BookCatalogEntry.model_validate(payload)
-
-
 @router.get("/api/books", response_model=List[schemas.Book])
 async def get_all_books(
     skip: int = 0,
@@ -229,27 +75,7 @@ async def get_book_catalog(
     sort_order: str = "asc",
     db: AsyncSession = Depends(get_db),
 ) -> List[schemas.BookCatalogEntry]:
-    books = await crud.get_book_catalog(db, q=q, sort_by=sort_by, sort_order=sort_order)
-    metadata_map = await _series_metadata_map(db, books)
-
-    series_books: dict[str, list[models.Book]] = {}
-    for book in books:
-        if book.series:
-            series_books.setdefault(book.series, []).append(book)
-
-    effective_series_tags: dict[str, list[str]] = {}
-    for series_name, group in series_books.items():
-        meta = metadata_map.get(series_name)
-        effective_series_tags[series_name] = crud.compute_effective_series_genre_tags(group, meta)
-
-    return [
-        _serialize_catalog_book(
-            book,
-            series_user_genre_tags=(metadata_map.get(book.series).user_genre_tags if book.series in metadata_map else []),
-            effective_series_genre_tags=effective_series_tags.get(book.series, []) if book.series else [],
-        )
-        for book in books
-    ]
+    return await build_book_catalog(db, q=q, sort_by=sort_by, sort_order=sort_order)
 
 
 @router.get("/api/series", response_model=List[str])
@@ -323,7 +149,7 @@ async def update_series_genres(
 
     crud.validate_genre_tags(body.user_genre_tags)
     canonical_name = books[0].series or series_name
-    user_genre_tags = _normalize_genre_tags(body.user_genre_tags)
+    user_genre_tags = normalize_genre_tags(body.user_genre_tags)
     metadata = await crud.set_series_user_genre_tags(
         db,
         series_name=canonical_name,
@@ -378,7 +204,7 @@ async def get_book_update_history(book_id: int, db: AsyncSession = Depends(get_d
         raise HTTPException(status_code=404, detail="Book not found")
 
     logs = await crud.get_book_logs(db, book_id)
-    return _build_chapter_update_history(book_id, logs)
+    return build_chapter_update_history(book_id, logs)
 
 
 @router.get("/api/books/{book_id}", response_model=schemas.Book)
@@ -401,7 +227,7 @@ async def update_book_details(
     previous_series = db_book.series
     previous_remote_ids = db_book.metadata_remote_ids
     if book_update.user_genre_tags is not None:
-        book_update.user_genre_tags = _normalize_genre_tags(book_update.user_genre_tags)
+        book_update.user_genre_tags = normalize_genre_tags(book_update.user_genre_tags)
     update_dict = book_update.model_dump(exclude_unset=True)
     updated_book = await crud.update_book(db=db, book=db_book, update_data=book_update)
     if "content_selectors" in update_dict or "removed_chapters" in update_dict:
