+++ frontend/src/components/BookSettings.jsx
@@ -1,26 +1,34 @@
-import { useState, useEffect, useRef } from "react";
+import { useState, useRef } from "react";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 
 import {
   deleteBook,
   detachBookSource,
-  getApiCoverUrl,
   getBook,
   getBookChapters,
   getBookCleanedChapters,
   getBookUpdateHistory,
-  getMatchedConfigs,
-  previewCleaning,
   processBook,
-  queueMetadataSync,
   refreshBook,
+  updateBook,
+} from "../api/books";
+import { getMatchedConfigs, previewCleaning } from "../api/cleaning";
+import {
+  getApiCoverUrl,
   retryBookCover,
   setBookCoverUrl,
-  updateBook,
   uploadBookCover,
-} from "../api/books";
+} from "../api/covers";
+import { queueMetadataSync } from "../api/metadata";
 import { getSeries } from "../api/series";
 import BookSettingsChapters from "./BookSettingsChapters";
+import {
+  ChapterUpdateHistory,
+  SelectorPills,
+  SourceTagList,
+  SyncedGenreTagList,
+} from "./book-settings/BookSettingsSections";
+import { useBookSettingsForm } from "./book-settings/useBookSettingsForm";
 
 const fetchChapters = async ({ queryKey }) => {
   const [_key, bookId] = queryKey;
@@ -41,258 +49,6 @@ const fetchUpdateHistory = async ({ queryKey }) => {
   const [_key, bookId] = queryKey;
   return getBookUpdateHistory(bookId);
 };
-
-const COMMON_REMOTE_ID_KEYS = [
-  "isbn_10",
-  "isbn_13",
-  "google_books_volume_id",
-  "open_library_work_key",
-  "open_library_edition_key",
-  "open_library_author_key",
-];
-
-function splitRemoteIds(remoteIds) {
-  const source = remoteIds && typeof remoteIds === "object" ? remoteIds : {};
-  const common = {};
-  COMMON_REMOTE_ID_KEYS.forEach((key) => {
-    common[key] = source[key] || "";
-  });
-
-  const extras = Object.fromEntries(
-    Object.entries(source).filter(
-      ([key]) => !COMMON_REMOTE_ID_KEYS.includes(key),
-    ),
-  );
-
-  return {
-    common,
-    extrasJson: Object.keys(extras).length
-      ? JSON.stringify(extras, null, 2)
-      : "",
-  };
-}
-
-function SelectorPills({ selectors, onChange }) {
-  const [inputValue, setInputValue] = useState("");
-
-  const addSelector = () => {
-    const trimmed = inputValue.trim();
-    if (trimmed && !selectors.includes(trimmed)) {
-      onChange([...selectors, trimmed]);
-    }
-    setInputValue("");
-  };
-
-  const removeSelector = (sel) => {
-    onChange(selectors.filter((s) => s !== sel));
-  };
-
-  return (
-    <div className="selector-pills">
-      <div className="pills">
-        {selectors.map((sel) => (
-          <span key={sel} className="pill">
-            {sel}
-            <button className="pill-remove" onClick={() => removeSelector(sel)}>
-              ×
-            </button>
-          </span>
-        ))}
-      </div>
-      <div className="pill-input">
-        <input
-          type="text"
-          placeholder="Add CSS selector, e.g. div.note"
-          value={inputValue}
-          onChange={(e) => setInputValue(e.target.value)}
-          onKeyDown={(e) => e.key === "Enter" && addSelector()}
-        />
-        <button onClick={addSelector}>Add</button>
-      </div>
-    </div>
-  );
-}
-
-function SourceTagList({ tags }) {
-  if (!tags?.length) return null;
-
-  return (
-    <div className="source-tag-list" aria-label="Source tags">
-      {tags.map((tag) => (
-        <span key={tag} className="source-tag">
-          {tag}
-        </span>
-      ))}
-    </div>
-  );
-}
-
-function SyncedGenreTagList({ tags }) {
-  if (!tags?.length) {
-    return (
-      <span className="settings-empty-value">No synced genre tags yet</span>
-    );
-  }
-
-  return (
-    <div className="genre-tag-list" aria-label="Synced genre tags">
-      {tags.map((tag) => (
-        <span key={tag} className="genre-tag">
-          {tag}
-        </span>
-      ))}
-    </div>
-  );
-}
-
-const EMPTY_UPDATE_HISTORY = {
-  history: [],
-  summary: {
-    total_update_events: 0,
-    total_chapters_added: 0,
-    total_words_added: 0,
-    average_words_per_week: null,
-    average_words_per_month: null,
-    average_days_between_updates: null,
-    predicted_next_update_at: null,
-    last_update_at: null,
-  },
-};
-
-function normalizeUpdateHistory(data) {
-  if (!data || !Array.isArray(data.history) || !data.summary) {
-    return EMPTY_UPDATE_HISTORY;
-  }
-  return data;
-}
-
-function formatNumber(value, options = {}) {
-  if (value == null || Number.isNaN(Number(value))) return "Not enough data";
-  return new Intl.NumberFormat(undefined, options).format(value);
-}
-
-function formatCompactNumber(value) {
-  if (value == null || Number.isNaN(Number(value))) return "0";
-  return new Intl.NumberFormat(undefined, {
-    notation: "compact",
-    maximumFractionDigits: 1,
-  }).format(value);
-}
-
-function formatDate(value) {
-  if (!value) return "Not enough data";
-  return new Date(value).toLocaleDateString(undefined, {
-    month: "short",
-    day: "numeric",
-    year: "numeric",
-  });
-}
-
-function formatHistoryEntryLabel(entry) {
-  if (entry.is_initial_sync) return "Initial sync";
-  if (entry.is_catch_up_sync) return "Catch-up sync";
-  return `+${entry.chapters_added} ch`;
-}
-
-function ChapterUpdateHistory({ updateHistory, isLoading, isError, error }) {
-  const { history, summary } = normalizeUpdateHistory(updateHistory);
-  const chartEntries = history.filter((entry) => entry.included_in_stats);
-  const maxWords = Math.max(
-    ...chartEntries.map((entry) => entry.words_added),
-    1,
-  );
-
-  return (
-    <section className="settings-section chapter-history-section">
-      <h3>Update History</h3>
-      {isLoading && <p className="hint">Loading update history...</p>}
-      {isError && (
-        <p className="error">
-          Update history failed: {error?.message || "Unable to load history"}
-        </p>
-      )}
-      {!isLoading && !isError && history.length === 0 && (
-        <p className="hint">
-          No tracked chapter updates yet. New chapter batches will appear here
-          after a refresh finds them.
-        </p>
-      )}
-      {!isLoading && !isError && history.length > 0 && (
-        <>
-          <div className="chapter-history-stats">
-            <div className="chapter-history-stat">
-              <span className="hint">Words / Week</span>
-              <strong>
-                {formatNumber(summary.average_words_per_week, {
-                  maximumFractionDigits: 0,
-                })}
-              </strong>
-            </div>
-            <div className="chapter-history-stat">
-              <span className="hint">Words / Month</span>
-              <strong>
-                {formatNumber(summary.average_words_per_month, {
-                  maximumFractionDigits: 0,
-                })}
-              </strong>
-            </div>
-            <div className="chapter-history-stat">
-              <span className="hint">Next Update</span>
-              <strong>{formatDate(summary.predicted_next_update_at)}</strong>
-            </div>
-          </div>
-
-          <div
-            className="chapter-history-chart"
-            aria-label="Words added by update date"
-          >
-            {chartEntries.length === 0 ? (
-              <p className="hint chapter-history-chart-empty">
-                No post-import chapter updates yet.
-              </p>
-            ) : (
-              chartEntries.slice(-12).map((entry) => {
-                const height = Math.max(
-                  (entry.words_added / maxWords) * 100,
-                  8,
-                );
-                return (
-                  <div className="chapter-history-bar-wrap" key={entry.id}>
-                    <div className="chapter-history-bar-stage">
-                      <div
-                        className="chapter-history-bar"
-                        style={{ height: `${height}%` }}
-                        title={`${formatDate(entry.timestamp)}: ${formatNumber(
-                          entry.words_added,
-                        )} words`}
-                      />
-                    </div>
-                    <span>{formatCompactNumber(entry.words_added)}</span>
-                  </div>
-                );
-              })
-            )}
-          </div>
-
-          <ul className="chapter-history-list">
-            {history
-              .slice()
-              .reverse()
-              .map((entry) => (
-                <li key={entry.id}>
-                  <span>{formatDate(entry.timestamp)}</span>
-                  <strong>
-                    {formatHistoryEntryLabel(entry)} ·{" "}
-                    {formatNumber(entry.words_added)} words
-                  </strong>
-                </li>
-              ))}
-          </ul>
-        </>
-      )}
-    </section>
-  );
-}
 
 function BookSettings({ book: initialBook, onBack }) {
   const queryClient = useQueryClient();
@@ -333,93 +89,51 @@ function BookSettings({ book: initialBook, onBack }) {
   // copy at the same deterministic URL.
   const [coverVersion, setCoverVersion] = useState(0);
 
-  const [title, setTitle] = useState(initialBook.title || "");
-  const [author, setAuthor] = useState(initialBook.author || "");
-  const [series, setSeries] = useState(initialBook.series || "");
-  const [seriesIndex, setSeriesIndex] = useState(
-    initialBook.series_index != null ? String(initialBook.series_index) : "",
-  );
-  const [notes, setNotes] = useState(initialBook.notes || "");
-  const [isbn10, setIsbn10] = useState(
-    initialBook.metadata_remote_ids?.isbn_10 || "",
-  );
-  const [isbn13, setIsbn13] = useState(
-    initialBook.metadata_remote_ids?.isbn_13 || "",
-  );
-  const [googleBooksVolumeId, setGoogleBooksVolumeId] = useState(
-    initialBook.metadata_remote_ids?.google_books_volume_id || "",
-  );
-  const [openLibraryWorkKey, setOpenLibraryWorkKey] = useState(
-    initialBook.metadata_remote_ids?.open_library_work_key || "",
-  );
-  const [openLibraryEditionKey, setOpenLibraryEditionKey] = useState(
-    initialBook.metadata_remote_ids?.open_library_edition_key || "",
-  );
-  const [openLibraryAuthorKey, setOpenLibraryAuthorKey] = useState(
-    initialBook.metadata_remote_ids?.open_library_author_key || "",
-  );
-  const [otherRemoteIdsJson, setOtherRemoteIdsJson] = useState(
-    splitRemoteIds(initialBook.metadata_remote_ids).extrasJson,
-  );
-  const [identifierError, setIdentifierError] = useState("");
-  const [userGenreTags, setUserGenreTags] = useState(
-    (initialBook.user_genre_tags || []).join(", "),
-  );
-  const [removedChapters, setRemovedChapters] = useState(
-    initialBook.removed_chapters || [],
-  );
-  const [contentSelectors, setContentSelectors] = useState(
-    initialBook.content_selectors || [],
-  );
-  const [previewResult, setPreviewResult] = useState(null);
+  const {
+    title,
+    setTitle,
+    author,
+    setAuthor,
+    series,
+    setSeries,
+    seriesIndex,
+    setSeriesIndex,
+    notes,
+    setNotes,
+    isbn10,
+    setIsbn10,
+    isbn13,
+    setIsbn13,
+    googleBooksVolumeId,
+    setGoogleBooksVolumeId,
+    openLibraryWorkKey,
+    setOpenLibraryWorkKey,
+    openLibraryEditionKey,
+    setOpenLibraryEditionKey,
+    openLibraryAuthorKey,
+    setOpenLibraryAuthorKey,
+    otherRemoteIdsJson,
+    setOtherRemoteIdsJson,
+    identifierError,
+    userGenreTags,
+    setUserGenreTags,
+    removedChapters,
+    setRemovedChapters,
+    contentSelectors,
+    setContentSelectors,
+    previewResult,
+    setPreviewResult,
+    chapterSearch,
+    setChapterSearch,
+    chaptersExpanded,
+    setChaptersExpanded,
+    chapterPreviewMode,
+    setChapterPreviewMode,
+    identifiersExpanded,
+    setIdentifiersExpanded,
+    getUpdatedFields,
+  } = useBookSettingsForm(initialBook);
   const [previewedChapter, setPreviewedChapter] = useState(null);
-  const [chapterSearch, setChapterSearch] = useState("");
-  const [chaptersExpanded, setChaptersExpanded] = useState(false);
-  const [chapterPreviewMode, setChapterPreviewMode] = useState("original");
-  const [identifiersExpanded, setIdentifiersExpanded] = useState(false);
-
-  // Reset form state when the parent navigates to a different book. We key this
-  // on the initial prop (not the polled `book`) so background refetches — e.g.
-  // while a refresh is in-flight — don't wipe out in-progress edits.
-  useEffect(() => {
-    setTitle(initialBook.title || "");
-    setAuthor(initialBook.author || "");
-    setSeries(initialBook.series || "");
-    setSeriesIndex(
-      initialBook.series_index != null ? String(initialBook.series_index) : "",
-    );
-    setNotes(initialBook.notes || "");
-    setIsbn10(initialBook.metadata_remote_ids?.isbn_10 || "");
-    setIsbn13(initialBook.metadata_remote_ids?.isbn_13 || "");
-    setGoogleBooksVolumeId(
-      initialBook.metadata_remote_ids?.google_books_volume_id || "",
-    );
-    setOpenLibraryWorkKey(
-      initialBook.metadata_remote_ids?.open_library_work_key || "",
-    );
-    setOpenLibraryEditionKey(
-      initialBook.metadata_remote_ids?.open_library_edition_key || "",
-    );
-    setOpenLibraryAuthorKey(
-      initialBook.metadata_remote_ids?.open_library_author_key || "",
-    );
-    setOtherRemoteIdsJson(
-      splitRemoteIds(initialBook.metadata_remote_ids).extrasJson,
-    );
-    setIdentifierError("");
-    setUserGenreTags((initialBook.user_genre_tags || []).join(", "));
-    setRemovedChapters(initialBook.removed_chapters || []);
-    setContentSelectors(initialBook.content_selectors || []);
-    setPreviewResult(null);
-    setChapterSearch("");
-    setChaptersExpanded(false);
-    setChapterPreviewMode("original");
-    setIdentifiersExpanded(false);
-  }, [initialBook]);
-
-  useEffect(() => {
-    setPreviewResult(null);
-  }, [contentSelectors, removedChapters]);
 
   const { data: chapters = [], isLoading: chaptersLoading } = useQuery({
     queryKey: ["chapters", book.id],
@@ -554,60 +268,6 @@ function BookSettings({ book: initialBook, onBack }) {
   const metadataSyncMutation = useMutation({
     mutationFn: () => queueMetadataSync([book.id], "manual"),
   });
-
-  const getUpdatedFields = () => {
-    let extraRemoteIds = {};
-    if (otherRemoteIdsJson.trim()) {
-      try {
-        const parsed = JSON.parse(otherRemoteIdsJson);
-        if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-          setIdentifierError("Other identifiers must be a JSON object.");
-          return null;
-        }
-        extraRemoteIds = parsed;
-      } catch {
-        setIdentifierError("Other identifiers must be valid JSON.");
-        return null;
-      }
-    }
-
-    setIdentifierError("");
-
-    const metadataRemoteIds = {
-      ...extraRemoteIds,
-      ...(isbn10.trim() ? { isbn_10: isbn10.trim() } : {}),
-      ...(isbn13.trim() ? { isbn_13: isbn13.trim() } : {}),
-      ...(googleBooksVolumeId.trim()
-        ? { google_books_volume_id: googleBooksVolumeId.trim() }
-        : {}),
-      ...(openLibraryWorkKey.trim()
-        ? { open_library_work_key: openLibraryWorkKey.trim() }
-        : {}),
-      ...(openLibraryEditionKey.trim()
-        ? { open_library_edition_key: openLibraryEditionKey.trim() }
-        : {}),
-      ...(openLibraryAuthorKey.trim()
-        ? { open_library_author_key: openLibraryAuthorKey.trim() }
-        : {}),
-    };
-
-    return {
-      title,
-      author,
-      series: series.trim() || null,
-      series_index: seriesIndex.trim() ? Number.parseFloat(seriesIndex) : null,
-      user_genre_tags: userGenreTags
-        .split(",")
-        .map((tag) => tag.trim())
-        .filter(Boolean),
-      metadata_remote_ids: Object.keys(metadataRemoteIds).length
-        ? metadataRemoteIds
-        : null,
-      removed_chapters: removedChapters,
-      content_selectors: contentSelectors,
-      notes: notes || null,
-    };
-  };
 
   const handleSave = () => {
     const payload = getUpdatedFields();
